+++ libopensimcreator/Graphics/OpenSimDecorationOptionFlags.h
@@ -22,9 +22,10 @@ namespace osc
         ShouldShowForceLinearComponent                      = 1<<8,
         ShouldShowForceAngularComponent                     = 1<<9,
         ShouldShowPointForces                               = 1<<10,
-        NUM_FLAGS                                           =    11,
+        ShouldShowScholz2015ObstacleContactHints            = 1<<11,
+        NUM_FLAGS                                           =    12,
 
-        Default = ShouldShowPointToPointSprings,
+        Default = ShouldShowPointToPointSprings | ShouldShowScholz2015ObstacleContactHints,
     };
     using OpenSimDecorationOptionFlags = Flags<OpenSimDecorationOptionFlag>;